+++ internal/rooms/delivery/handler.go
@@ -2,9 +2,9 @@ package delivery
 
 import (
 	"fmt"
+	"net/http"
 	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
-	"net/http"
 	"zoomer/internal/auth/repository"
 	"zoomer/internal/models"
 	"zoomer/internal/rooms/presenter"